add linegen.py
